+++ internal/app/types/user_mongo.go
@@ -41,14 +41,6 @@ type User struct {
 	LastNotificationSentDate time.Time `json:"lastNotificationSentDate,omitempty" bson:"lastNotificationSentDate,omitempty"`
 }
 
-// UserESRecord is the data that will store into the elastic search.
-type UserESRecord struct {
-	UserID    string `json:"userID"`
-	FirstName string `json:"firstName,omitempty"`
-	LastName  string `json:"lastName,omitempty"`
-	Email     string `json:"email,omitempty"`
-}
-
 // Helper types
 
 type FindUserResult struct {
